Put all filters in one separate folder from the rest of the JS
