+++ javascript/filters/type_filter.js
@@ -1,4 +1,4 @@
-export function filterType() {
+export function filterType(data) {
     const onlineChallenges = document.querySelector("#online-challenges");
     const onsiteChallenges = document.querySelector("#on-site-challenges");
